--- HellDivers_UnityProject/Assets/ProCore.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f6297a6b7f646704b9257e15afa73a1b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
