+++ HellDivers_UnityProject/Assets/Scripts/Player/PlayerController.cs
@@ -4,83 +4,34 @@ using UnityEngine;
 
 public class PlayerController : MonoBehaviour
 {
-    public float Speed { get { return m_Speed; } set { m_Speed = value; } }
-
-    [SerializeField] private float m_Speed = 5f;
-
     #region Private Variable
-
     private CharacterController m_Controller;
     private Transform m_Cam;
-    private Vector3 m_CamFoward;
-    private Vector3 m_Direction;
+    private Vector3 m_CamForward;
     private Vector3 m_Move;
+    private Vector3 m_Direction;
     private Ray m_MouseRay;
     private RaycastHit m_MouseHit;
-    private bool bChange;
+    private PlayerAnimationsContorller PAC;
     #endregion Private Variable
+
+    private bool bRun = false;
+    private bool bInBattle = false;
+    private bool bAttack = false;
 
     private void Start()
     {
+        PAC = this.GetComponent<PlayerAnimationsContorller>();
         m_Controller = this.GetComponent<CharacterController>();
 
         if (Camera.main != null)
         {
             m_Cam = Camera.main.transform;
         }
-        bChange = true;
     }
-
-    private void Update()
+    private void FixedUpdate()
     {
-        if (Input.GetButton("Horizontal") || Input.GetButton("Vertical"))
-        {
-            Move();
-        }
-        else
-        {
-            PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_IDLE;
-        }
-        if (m_Controller.isGrounded == false)
-        {
-            m_Controller.Move(Physics.gravity * Time.deltaTime);
-        }
-        if (Input.GetMouseButton(0) || Input.GetMouseButton(1))
-        {
-            FaceDirection();
-            Attack();
-        }
-        #region DisplayAnimation
-        if (PlayerAnimationsContorller.m_AttackState != ePlayerAttack.ANI_THROWSTANDBY)
-        {
-            if ((Input.GetKeyDown(KeyCode.Alpha1) || Input.GetKeyDown(KeyCode.Alpha2)))
-            {
-                PlayerAnimationsContorller.m_AttackState = ePlayerAttack.ANI_SWITCHWEAPON;
-            }
-            if (Input.GetKeyDown(KeyCode.Alpha4))
-            {
-                PlayerAnimationsContorller.m_AttackState = ePlayerAttack.ANI_THROW;
-                Debug.Log("Press");
-            }
-            if (Input.GetKeyDown(KeyCode.R))
-            {
-                PlayerAnimationsContorller.m_AttackState = ePlayerAttack.ANI_RELOAD;
-            }
-            if (Input.GetKeyDown(KeyCode.Y))
-            {
-                PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_DEATH;
-            }
-        }
-
-        if (Input.GetKeyDown(KeyCode.Space))
-        {
-            PlayerAnimationsContorller.m_AnyState = ePlayerAnyState.ANI_ROLL;
-        }
-        if (Input.GetKeyDown(KeyCode.T))
-        {
-            PlayerAnimationsContorller.m_AnyState = ePlayerAnyState.ANI_GETHURT;
-        }
-        #endregion DisplayAnimation
+        Move();
     }
 
     private void Move()
@@ -88,107 +39,52 @@ public class PlayerController : MonoBehaviour
         float h = Input.GetAxis("Horizontal");
         float v = Input.GetAxis("Vertical");
 
+
         if (m_Cam != null)
         {
-            m_CamFoward = Vector3.Scale(m_Cam.forward, Vector3.forward + Vector3.right);
-            m_Direction = m_CamFoward.normalized * v + m_Cam.right * h;
+            m_CamForward = Vector3.Scale(m_Cam.forward, new Vector3(1, 0, 1)).normalized;
+            m_Move = v * m_CamForward + h * m_Cam.right;
         }
         else
         {
-            m_Direction = Vector3.forward * v + Vector3.right * h;
+            m_Move = v * Vector3.forward + h * Vector3.right;
         }
 
-        if (m_Direction.magnitude > 1)
-        {
-            m_Direction.Normalize();
-        }
+        bRun = (Input.GetButton("Run")) ? true : false;
 
-        m_Move = m_Direction * m_Speed * Time.deltaTime;
+        if (m_Move.magnitude > 1) m_Move.Normalize();
 
-        float fAngle = Vector3.Angle(this.transform.forward, m_Direction.normalized);
-        float dotTurnRight = Vector3.Dot(this.transform.right, m_Direction.normalized);
-
-        if (fAngle > 80)
-        {
-            if (dotTurnRight >= 0)
-            {
-                //PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_TURNRIGHT90;
-                Debug.Log("Turn Right");
-            }
-            else if (dotTurnRight < 0)
-            {
-                //PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_TURNLEFT90;
-                Debug.Log("Turn Left");
-            }
-
-        }
         if (Input.GetMouseButton(0) || Input.GetMouseButton(1))
         {
-            PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_WALKSHOOT;
-        }
-        else
-        {
-            PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_WALK;
-            if (Input.GetButton("Run"))
+            FaceDirection();
+            bInBattle = true;
+            if (Input.GetMouseButtonDown(0))
             {
-                m_Move *= 3.0f;
-                PlayerAnimationsContorller.m_MoveState = ePlayerAnimationState.ANI_RUN;
+                bAttack = true;
             }
         }
+        else bInBattle = false;
 
-
-
-        if (m_Controller.isGrounded == false)
-        {
-            m_Move += Physics.gravity * Time.deltaTime;
-        }
-
-
-        if (fAngle < 80)
-        {
-        }
-            this.transform.forward = m_Direction;
-            m_Controller.Move(m_Move);
-
-       
+        PAC.Move(m_Move, m_Direction, bRun, bInBattle, bAttack);
     }
 
     private void FaceDirection()
     {
         float vHigh = Camera.main.transform.position.y - this.transform.position.y;
-
         Ray MouseRay = Camera.main.ScreenPointToRay(Input.mousePosition);
         Vector3 vMouseRay = MouseRay.direction;
-
         vMouseRay.Normalize();
-        float angle = Vector3.Dot(new Vector3(0, -1, 0), vMouseRay);
 
+        float angle = Vector3.Dot(new Vector3(0, -1, 0), vMouseRay);
         float distance = vHigh / angle;
         Vector3 endPoint = MouseRay.GetPoint(distance);
+
         m_Direction = endPoint - this.transform.position;
         m_Direction.y = 0.0f;
-
         if (m_Direction.magnitude < 0.1f) return;
-        m_Controller.transform.forward = m_Direction;
+        if (m_Direction.magnitude > 1) m_Direction.Normalize();
     }
-
-    private void Attack()
-    {
-        if (Input.GetMouseButton(0))
-        {
-            if (PlayerAnimationsContorller.m_AttackState == ePlayerAttack.ANI_GUNPLAY)
-            {
-                PlayerAnimationsContorller.m_AttackState = ePlayerAttack.ANI_SHOOT;
-            }
-            if (PlayerAnimationsContorller.m_AttackState == ePlayerAttack.ANI_THROWSTANDBY)
-            {
-                PlayerAnimationsContorller.m_AttackState = ePlayerAttack.ANI_THROWOUT;
-            }
-        }
-    }
-
 #if UNITY_EDITOR
-
     private void OnDrawGizmos()
     {
         Gizmos.color = Color.cyan;
