--- HellDivers_UnityProject/Assets/Test/Woody/Maps.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bca25427f53fbdd439ae26ba1d443d53
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
