--- HellDivers_UnityProject/Assets/Test/Woody/New Animator Controller.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 135717fcb09902e4fb5dd133b053abbe
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
