--- HellDivers_UnityProject/Assets/Test/Woody/PlayerControl.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 11e25f10645de0946a54845de80a3112
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
