--- HellDivers_UnityProject/Assets/Test/Woody/Resources/Map1.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b1077886c6ed7bd4cb2066bdb19ab7ea
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
