--- HellDivers_UnityProject/Assets/Test/Woody/Resources/Map2.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3e683ef33a869f641b177207b9e5a3a8
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
