--- HellDivers_UnityProject/Assets/Test/Woody/Resources/Map3.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bedf7ae5476be0f48b60749a6c765f0a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
