--- HellDivers_UnityProject/Assets/Test/Woody/Resources/Map4.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f11858324c9fd5c4ba63b2ebd20d4a7b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
