--- HellDivers_UnityProject/Assets/Test/Woody/scripts/AniTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a1c2a60ddeec41d41bd547ba9bfa8880
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
