--- HellDivers_UnityProject/Assets/Test/Woody/scripts/ControllerTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3a9374bb2c961504a8f335573146c145
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
